requirements.txt
Flask
flask-socketio
eventlet
tensorflow>=2.6
numpy
opencv-python
pyserial
